Move shorthand processing up and address comments
- Move shorthand processing before NOT processing in normalization.
Thus making shorthand processing an individual unit of work in
the normalization process instead of a sub-unit of NOT processing as it
was previously.
- Add comment and made variable name changes to clarify the notion of
what shorthand and longhand notion is w.r.t. the query normalization.
- Add comment for NOT operator compression functionality
- Clarified comments and moved the comments when appropriate throughout
the Validator.
- Add comment and made variable name changes to clarify the handling
of multiple, sequential NOT operator during query normalization.
- Use loop construct in tests to build query map that contains
nested, sequential $not operators.